refactor(evm): move routemesh to presets and consolidate rpc internals
- Move routemesh module from rpc/ to presets/ (logically belongs with transports/layers)
- Remove rpc/internal/index.ts, export internals directly from rpc/index.ts
- Update imports across affected files
