--- evm/src/rpc/internal/index.ts
@@ -1,2 +1,0 @@
-export { makeRpcCacheLive, RpcCache } from "../cache.js";
-export { RequestDedup, RequestDedupLive } from "../dedup.js";
